Added rct class
Default recurrent classes:
rct,
rctBpg

Allows for multiple training sets to be run through quicker on a single model by synchronously evaluating multiple predictions on specified inputs.

prep(int a): The maximum number of instances of the template model that are intended to be used at once should be prepared (instantiated).
unroll(int a): The number of instances of the template model that are intended to be trained on at any certain time.