Revoked ptr<vector> in tensor
ptr<vector<tensor>> found no use in the framework, and the accepted method of tensor linking could not provide a use for ptr<vector<tensor>>.

Previously, it was accepted that recurrent networks would be resizing their x and y at unroll time, but this would cause complications when compiling a tensor to an individual element of the variable length tensor.

E.g.
Take tensor A, and B.
A is a tensor of rank 2.
B is a tensor of rank 1, and is sharing source material with row index 9 of A.
Assume that A resizes to size 9 from 10. The rows that were previously linked to by B would no longer be a part of A, and thus the two would no longer share what they once did.

Resizing is still an option with tensors.
However, because of the flaw mentioned above, it will not be incoorporated majorly in the modeling side of the framework outside of construction time, or preparation time.

Tensor functionality remains similar.

Recurrent models' x & y will now appear invariable in length due to their rescriction from resizing at unroll time.

Thus, the models' 2nd order x and y are to be regarded by the user as needing to ignore certain 1st order elements, defined by how many timesteps are unrolled.