Added Attention LSTM
Attention LSTM Prep and Unroll Methods are nonintuitive at a shallow glance.
prep(a_a, a_b):
a_a does not imply the rank 2 length of the input matrix. Instead, a_b implies the rank 2 length of the input matrix. a_a is the argument for the number of output timesteps.

... (simply put)
a_a should be the length of y.
a_b should be the length of x.