Fixed Convergence of NTM
The NTM model had several issues, preventing convergence.
The first was improper tensor linkage in the ntm_wh model. The error was specifically that 'x' was not linked to the a_model or e_model, and so those models would always receive a tensor full of zeros as their inputs.
Secondly, ntm_reader and ntm_writer suffered from improper tensor linkage as well. both models had their "mx_grad" linked directly to "internal_addresser->x_grad". The problem here is that "mx_grad" would be overridden when the "internal_addresser" carried backward, erasing any gradient that was once in that tensor. The link has been removed in the compile methods, and the tensor "mx_grad" is now a linear accumulation of both "internal_addresser->x_grad" and the gradient calculated manually by the ntm_reader.